Move and improve MultiSourceTaskNamesCondition
